ref: Makes some of the JingleOfferFactory methods non-static.
Video content creation is parametrizable via some configuration
options, so these methods need access to the ConfigurationService
that is discovered and initialized in the FocusBundleActivator start.

After the ConfigurationService is initialized, we initialize the
JingleOfferFactory and we also make it a static member of the
FocusBundleActivator, just like the ConfigurationService.