ref: deprecate Recorder
JibriRecorder will no longer extend deprecated Recorder. Also
JibriRecorder instance lifetime will no longer be related to
the lifetime of a Colibri conference on the JVB. That's because it needs
to report Jibri availability even if the conference has been stopped.
Like for example when there's just one participant left and the JVB
conference is being expired.