fix: Fixes a bug with Participant-s
being created before we have valid presence for them.

This removes the presence cache, and moves the logic that handled a new
MUC member from MemberListener#joined to the presence listener. This
ensures that when we create a new Participant, we have already processed
the ChatRoomMember's presence and read extracted the information that
we need (e.g. region info, or in the future statistics ID).
MemberListener#joined is now just ignored.