Do not reuse HTTP client connections during reservation
When Jicofo sends HTTP requests to the reservation system
using the reservation API, the request would fail with
NoHttpResponseException for roughly every even numbered
request. e.g. When creating conference one after another,
the second conference creation would fail.

We have pin point the cause of the error to incorrect
usage of the Apache HTTP client by Jicofo. This patch
is an attempt to work around the bug by not reusing
a HTTP connection.