fix(SSRCValidator): getIndependentFidGroups
It used to work when the SSRCs were being verified only
on the per participant basis, but it turned out there is
a bug when the conference sources have to be tested as
a whole.

The problem was that FID groups would be marked
as independent if there are more than 1 SIM groups in
a case where FID group does not belong to the first SIM
group, but does belong to the second or another one.