Add option to always trust XMPP domain certificate.
When jicofo connects to the XMPP domain, it uses a default, Smack-provided
implementation that controls verification and validation of the server-provided
TLS certificate. In certain setups, it is desirable to override this behavior,
and allow any certificate that's provided.

This commit adds an option that, once enabled, causes Jicofo to trust any and
all certificate that is provided by the XMPP server that it conntects to.