Another iteration of the Octo support (#262)
* ref: Renames endpointName to endpointId.

* ref: Refactors ChannelAllocator.

* ref: Extracts an AbstractParticipant class,

which is to be used for the bridge-to-bridge (Octo) channels.

* fix: Fixes a bug in group comparison.

* fix: Fixes a bug in SourceGroup.equals()

The previous implementation would not take the order of the sources into
account, and would only check if the group's sources is a subset of the
other group's sources. E.g. groups with the following sources would be
considered equal: [1, 2] and [2, 1, 3]

* ref: Clean up code, simplifies code (using java8 features),

uses CopyOnWriteArrayList consistently.

* ref: Renames fields and methods for consistency.

Uses "participant" instead of "peer" in JitsiMeetConferenceImpl when
referring to Participants, since they are not peers (except in the
context of a Jingle session, which is not how it was used here).

participant -> member or occupant (as appropriate) in ChatRoomImpl
contact address -> Jid in ChatRoom
roomName -> roomJid in ChatRoom
mucJid -> occupantJid (since XEP-0045 uses this term)

* ref: Adds Participant#isSessionEstablished(),

simplifies code.

* minor: updates docs and formatting.

* ref: Adapts to changes in jitsi.

* ref: Returns a more specific class from getParticipant().

* feat: Adds SplitBridgeSelectionStrategy

Only used for testing.

* feat: Supports multiple bridges with Octo.

* fix: Avoids using ColibriConferenceImpl directly.

* minor: Formatting fixes.

* mvn: Updates jitsi-desktop.version.

* minor: Formatting fixes.

* ref: Renames BridgeState to Bridge.

* fix: Removes unnecessary checks for 'canceled', makes a method abstract.