Optimizations around participant re-invite and failure handling (#334)
* feat(Participant): add terminateBridgeSession

Adds the 'terminateBridgeSession' method to Participant which will
remove it from the current BridgeSession and cleanup the fields related
to the session. Those include the channel allocator's thread, transport
info (ICE candidates and DTLS fingerprint) and Colibri channels info
(the channels are expired when BridgeSession is terminated).

* ref(JitsiMeetConferenceImpl): remove 'onChannelAllocationFailed'

The whole logic inside of 'onChannelAllocationFailed' makes no sense,
because the 'onBridgeDown' method is called earlier in the call chain of
AbstractChannelAllocator from the 'allocateChannels' method.

The 'onBridgeDown' contains very similar if not effectively the same
re-invite logic to the one from the method removed.