fix: channel re-allocation race condition
Adds the option to synchronously expire Colibri channels with blocking
the current thread until the RESULT packet is received from the Bridge.

In the ICE failed/re-invite scenario Jicofo used to send expire and
allocate Colibri requests almost simultaneously. It can happen that
the bridge would be processing them both at the same time which can
result in IceTransportManager not being re-created. In that case ICE
ufrag and pwd attributes would remain the same and there would be no ICE
restart on the client side.