fix: Octo signaling when endpoints join muted.
Injects a random SSRC for endpoints which do not advertise one in
session-accept. Endpoints which join a conference audio muted and
video muted do not advertise an SSRC. Without any SSRCs, remote
bridges are not aware of these endpoints, leading to problems with
data channel messages and dominant speaker identification.