fix the shibboleth web handler
since the servlet has a path spec of '/*', it matches all urls and, if
it doesn't end up handling a request, that request is still marked as
'handled' so it doesn't fall through to the next handler.

now, we'll move the shibboleth handler first but make sure it handles
only the shibboleth urls: all others will not be marked as handled and
will fall through to the next handler.