fix: remove sources from participants broken by b70696ec
..and add unit test which include source removal validation.

Adjust the MediaSourceMap to not accept duplicates. It is used by
MockParticipant to add sources and it should mimic current Jitsi Meet
client's behavior of not adding duplicated sources.

Jicofo is not 100% accurate with tracking unadvertised sources.
Some sources that were included in session-initiate are re-advertised
when session-accept is received from the client, because
they were scheduled for addition while the jingle session is not active
yet. A check on the client is easier(and it's been there forever) rather
than trying to achieve 100% accuracy in the current implementation.