Propagate sources upon ice restart + some. (#491)
* fix: Propagate source and source-groups on transport-accept.

A participant sends a transport-accept when he's re-invited by the
focus. When that happens, the participant may land on a different
bridge. The new bridge needs to be informed about the sources and the
source-groups from the participant.

* fix: Remove sources of a terminated participant from Octo.

* ref(Jingle): remove isAccepted flag

It was a minor optimization to expire channels
faster in case the client doesn't ack the invite.
It's probably better to ignore timeout and let
the channels expire on the bridge due to no activity.
This will give the client more time in case there
are connectivity issues.

Co-authored-by: paweldomas <pawel.domas@jitsi.org>
Co-authored-by: Boris Grozev <boris@jitsi.org>