ref: remove sync channel expire
Doing synchronous expire blocks Smack's Single Threaded
executor and builds up a delay on the processing queue.
It also may time out other jobs that are waiting for
a packet to arrive if the expire will block and timeout.

It's recommended to run the JVB2 with the IQ_HANDLER_MODE
set to 'sync' in the MUC configuration to have ICE restarts
working.