feat: process 'session-terminate'
Implements 'session-terminate' which allows the clients to end the session.
If there's the restart attribute set to true in the <bridge-session> extension,
then Jicofo will start a new session after tearing down the existing one.

Adds rate limiting on the restart requests to prevent abuse and/or client bugs.
There must be at least 10 second gap between the requests and no more than
3 requests are allowed in the last minute.