Octo user migration fixes (#617)
This fixes the following case of moving participants to another bridge after a failure:

Endpoints A and B on jvb1, endpoints C and D on jvb2. Kill jvb1. Endpoints A and B are moved to jvb3. The following happens:

Video between C and D continues to work, but not for any other endpoints.
The jicofo logs seem normal — it sees jvb1 go away, allocates channels on a new bridge, sends transport-replace, receives transport-accept.
On jvb2 things seem normal — remove relay for jvb1, add relay for jvb3.
jvb3 receives a conference request and creates OctoEndpoint-s for endpoints A and B. Then it replaces that with local Endpoint instances and gathers candidates. While in that state, jvb3 spams the logs with Received a message with an unexpected type: EndpointConnectivityStatusChangeEvent. (still investigating)
This PR fixes several issues:

The first problem is the weird transition from octo endpoints to local endpoint again on jvb3. This is fixed in a888ec7. The fix prevents the inclusion of the endpoints of the crashed bridge (jvb1) in the octo channel allocation on the bridge that was chosen to replace the crashed bridge (jvb3).
The second problem is that after a re-invite, there's no source/source group propagation, preventing media to flow back to the local participants of the old healthy bridge. This is fixed in e9f1c3b