fix: Avoid a race in Smack 4.4.3 which causes
a MUC to be unusable and Jicofo thread leaks. Handling the self-presence
response races with leave(), resulting in an entry in the occupants map.
This entry prevents future joins as initial self-presence is not
recognized and therefore not processed.

Not sending and updated BRIDGE_COUNT before leaving makes the race
significantly less likely, but doesn't fix the underlying problem.