fix: Fix broken Colibri2SessionManager state.
When a session is removed the participantsBySession map is left with an
entry for the session (with an empty list of participants). This results
in the bridge being passed to BridgeSelector indefinitely.

In particular, this causes a serious problem when the bridge does not
have a relay, because BridgeSelector returns it back immediately, even
if the instance has disconnected.