fix: Temporary fix for JingleSession-s being GC too early
JingleSessions register with JingleIqHandler which saves a weak
reference and there are no other references to the object until
the session is accepted and saved as Participant.jingleSession. If
the weak reference expires before session-accept is received the
session fails to establish.

This is a quick and dirty fix, a proper solution will follow later.