fix: Fix ChatRoomImpl log level. (#1259)
LoggerImpl delegates to a java.util.logging.Logger and setting the level
is also delegated. As a result, the shared delegate's log level is
changed every time. We should probably change that logger behavior to
only change the level of each instance, but I'm not convinced that it's
safe to do so.